+++ wires-core/wires-core-scratchpad/src/main/java/org/kie/wires/core/scratchpad/client/factories/dynamic/RectangleFactory.java
@@ -69,12 +69,12 @@ public class RectangleFactory extends AbstractBaseFactory<Rectangle> {
     }
 
     @Override
-    protected int getWidth() {
+    protected double getWidth() {
         return SHAPE_SIZE_X + ( ShapesUtils.RGB_STROKE_WIDTH_SHAPE * 2 );
     }
 
     @Override
-    protected int getHeight() {
+    protected double getHeight() {
         return SHAPE_SIZE_Y + ( ShapesUtils.RGB_STROKE_WIDTH_SHAPE * 2 );
     }
